pyproject.toml
[tool.poetry]
package-mode = false
name = "getdocs"
version = "0.1.0"
description = ""
authors = ["Aliaksey Khrystosay <stroitop214@gmail.com>"]
license = "ITM"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.30"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
celery = "^5.4.0"
redis = "^5.0.4"
pytesseract = "^0.3.10"
pytest = "^8.2.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.7"
anyio = "^4.4.0"
pip = "^24.1"
pytest-dotenv = "^0.5.2"
imageio = "^2.34.2"
pytest-env = "^1.1.3"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"